refactor: tighten sidecar diagnostics cleanup
