+++ packages/frontend/src/components/features/diagnostics/diagnostics-panel-model.test.ts
@@ -1,111 +1,13 @@
 import { describe, expect, test } from "bun:test";
-import type {
-  BeadsCheck,
-  RuntimeDescriptor,
-  RuntimeInstanceSummary,
-  WorkspaceRecord,
-} from "@openducktor/contracts";
 import { CODEX_RUNTIME_DESCRIPTOR, OPENCODE_RUNTIME_DESCRIPTOR } from "@openducktor/contracts";
-import {
-  type BeadsCheckFixtureOverrides,
-  createBeadsCheckFixture,
-} from "@/test-utils/shared-test-fixtures";
-import type { RepoRuntimeHealthCheck } from "@/types/diagnostics";
 import { buildDiagnosticsPanelModel } from "./diagnostics-panel-model";
-
-const runtimeDefinitions: RuntimeDescriptor[] = [OPENCODE_RUNTIME_DESCRIPTOR];
-
-const runtimeSummary: RuntimeInstanceSummary = {
-  kind: "opencode",
-  runtimeId: "runtime-1",
-  repoPath: "/repo",
-  taskId: null,
-  role: "workspace",
-  workingDirectory: "/repo",
-  runtimeRoute: {
-    type: "local_http",
-    endpoint: "http://127.0.0.1:49700",
-  },
-  startedAt: "2026-02-20T12:00:00.000Z",
-  descriptor: OPENCODE_RUNTIME_DESCRIPTOR,
-};
-
-type RepoHealthOverrides = Omit<Partial<RepoRuntimeHealthCheck>, "runtime" | "mcp"> & {
-  runtime?: Partial<RepoRuntimeHealthCheck["runtime"]>;
-  mcp?: Partial<NonNullable<RepoRuntimeHealthCheck["mcp"]>>;
-};
-
-const makeRepoHealth = (overrides: RepoHealthOverrides = {}): RepoRuntimeHealthCheck => {
-  const checkedAt = overrides.checkedAt ?? "2026-02-20T12:01:00.000Z";
-  const runtime: RepoRuntimeHealthCheck["runtime"] = {
-    status: "ready",
-    stage: "runtime_ready",
-    observation: null,
-    instance: null,
-    startedAt: null,
-    updatedAt: checkedAt,
-    elapsedMs: null,
-    attempts: null,
-    detail: null,
-    failureKind: null,
-    failureReason: null,
-    ...overrides.runtime,
-  };
-  const mcp: NonNullable<RepoRuntimeHealthCheck["mcp"]> = {
-    supported: true,
-    status: "connected",
-    serverName: "openducktor",
-    serverStatus: "connected",
-    toolIds: [],
-    detail: null,
-    failureKind: null,
-    ...overrides.mcp,
-  };
-
-  return {
-    status:
-      overrides.status ??
-      (runtime.status === "error" || mcp.status === "error"
-        ? "error"
-        : mcp.status === "checking" ||
-            mcp.status === "reconnecting" ||
-            mcp.status === "waiting_for_runtime"
-          ? "checking"
-          : runtime.status),
-    checkedAt,
-    runtime,
-    mcp,
-  };
-};
-
-const makeBeadsCheck = (overrides: BeadsCheckFixtureOverrides = {}): BeadsCheck =>
-  createBeadsCheckFixture(
-    {
-      beadsPath: "/Users/dev/.openducktor/beads/repo/.beads",
-      repoStoreHealth: {
-        attachment: {
-          path: "/Users/dev/.openducktor/beads/repo/.beads",
-          databaseName: "repo_db",
-        },
-      },
-    },
-    overrides,
-  );
-
-const makeWorkspace = (
-  repoPath: string,
-  overrides: Partial<WorkspaceRecord> = {},
-): WorkspaceRecord => ({
-  workspaceId: repoPath.split("/").filter(Boolean).at(-1) ?? "repo",
-  workspaceName: repoPath.split("/").filter(Boolean).at(-1) ?? "repo",
-  repoPath,
-  isActive: true,
-  hasConfig: true,
-  configuredWorktreeBasePath: "/worktrees",
-  defaultWorktreeBasePath: "/Users/dev/.openducktor/worktrees/repo",
-  effectiveWorktreeBasePath: "/worktrees",
-  ...overrides,
-});
+import {
+  makeBeadsCheck,
+  makeRepoHealth,
+  makeWorkspace,
+  runtimeDefinitions,
+  runtimeSummary,
+} from "./diagnostics-panel-model-test-fixtures";
 
 describe("buildDiagnosticsPanelModel", () => {
   test("returns no-repository summary and empty-state messages when no repository is selected", () => {
@@ -396,143 +298,6 @@ describe("buildDiagnosticsPanelModel", () => {
     expect(beadsSection?.errors).toEqual([
       "Shared Dolt database repo_db is missing and restore is required",
     ]);
-  });
-
-  test("renders bundled and override executable provenance for Beads diagnostics", () => {
-    const model = buildDiagnosticsPanelModel({
-      workspaceRepoPath: "/repo",
-      activeWorkspace: makeWorkspace("/repo"),
-      runtimeDefinitions,
-      isLoadingRuntimeDefinitions: false,
-      runtimeDefinitionsError: null,
-      runtimeCheck: {
-        gitOk: true,
-        gitVersion: "git version 2.50.1",
-        ghOk: true,
-        ghVersion: "gh version 2.73.0",
-        ghAuthOk: true,
-        ghAuthLogin: "octocat",
-        ghAuthError: null,
-        runtimes: [{ kind: "opencode", ok: true, version: "1.2.9" }],
-        errors: [],
-      },
-      beadsCheck: makeBeadsCheck({
-        beadsExecutable: {
-          displayLabel: "Bundled with OpenDucktor",
-          error: null,
-          path: "/Applications/OpenDucktor.app/Contents/Resources/bin/bd",
-          sourceCategory: "bundled_electron_resource",
-        },
-        doltExecutable: {
-          displayLabel: "Environment override",
-          error: null,
-          path: "/opt/dolt/bin/dolt",
-          sourceCategory: "environment_override",
-        },
-      }),
-      runtimeCheckFailureKind: null,
-      beadsCheckFailureKind: null,
-      runtimeHealthByRuntime: {
-        opencode: makeRepoHealth({
-          runtime: { instance: runtimeSummary },
-          mcp: { toolIds: [] },
-        }),
-      },
-      isLoadingChecks: false,
-    });
-
-    const beadsSection = model.sections.find((section) => section.key === "beads-store");
-
-    expect(beadsSection?.rows).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({
-          label: "Beads executable source",
-          value: "Bundled with OpenDucktor",
-        }),
-        expect.objectContaining({
-          label: "Beads executable path",
-          value: "/Applications/OpenDucktor.app/Contents/Resources/bin/bd",
-        }),
-        expect.objectContaining({ label: "Dolt executable source", value: "Environment override" }),
-        expect.objectContaining({ label: "Dolt executable path", value: "/opt/dolt/bin/dolt" }),
-      ]),
-    );
-  });
-
-  test("renders unavailable executable provenance without duplicating primary errors", () => {
-    const missingBeads =
-      "Packaged Electron Beads sidecar is missing or invalid: expected bd in /Applications/OpenDucktor.app/Contents/Resources/bin. This is an OpenDucktor packaging defect.";
-    const missingDolt =
-      "Packaged Electron Dolt sidecar is missing or invalid: expected dolt in /Applications/OpenDucktor.app/Contents/Resources/bin. This is an OpenDucktor packaging defect.";
-    const model = buildDiagnosticsPanelModel({
-      workspaceRepoPath: "/repo",
-      activeWorkspace: makeWorkspace("/repo"),
-      runtimeDefinitions,
-      isLoadingRuntimeDefinitions: false,
-      runtimeDefinitionsError: null,
-      runtimeCheck: {
-        gitOk: true,
-        gitVersion: "git version 2.50.1",
-        ghOk: true,
-        ghVersion: "gh version 2.73.0",
-        ghAuthOk: true,
-        ghAuthLogin: "octocat",
-        ghAuthError: null,
-        runtimes: [{ kind: "opencode", ok: true, version: "1.2.9" }],
-        errors: [],
-      },
-      beadsCheck: makeBeadsCheck({
-        beadsOk: false,
-        beadsPath: null,
-        beadsError: missingBeads,
-        beadsExecutable: {
-          displayLabel: "Unavailable",
-          error: missingBeads,
-          path: null,
-          sourceCategory: "unavailable",
-        },
-        doltExecutable: {
-          displayLabel: "Unavailable",
-          error: missingDolt,
-          path: null,
-          sourceCategory: "unavailable",
-        },
-        repoStoreHealth: {
-          category: "attachment_verification_failed",
-          status: "blocking",
-          isReady: false,
-          detail: missingBeads,
-          attachment: {
-            path: null,
-            databaseName: null,
-          },
-          sharedServer: {
-            host: null,
-            port: null,
-            ownershipState: "unavailable",
-          },
-        },
-      }),
-      runtimeCheckFailureKind: null,
-      beadsCheckFailureKind: null,
-      runtimeHealthByRuntime: {
-        opencode: makeRepoHealth({
-          runtime: { instance: runtimeSummary },
-          mcp: { toolIds: [] },
-        }),
-      },
-      isLoadingChecks: false,
-    });
-
-    const beadsSection = model.sections.find((section) => section.key === "beads-store");
-
-    expect(beadsSection?.rows).toEqual(
-      expect.arrayContaining([
-        expect.objectContaining({ label: "Beads executable path", value: "Unavailable" }),
-        expect.objectContaining({ label: "Dolt executable path", value: "Unavailable" }),
-      ]),
-    );
-    expect(beadsSection?.errors).toEqual([missingBeads, missingDolt]);
   });
 
   test("treats repositories using the default worktree path as healthy", () => {
